update 404 and sidebar
