+++ src/app/views/layout/ad-sidebar/ad-sidebar.component.html
@@ -1,1 +1,6 @@
-<p>ad-sidebar works!</p>
+
+<div class="add_part">
+    <h2 class="addSection">Ads Section</h2>
+    <hr class="d-none d-md-block my-2 ms-0" />
+    <img class="ad_img" src="assets\images\keralaspices-ad300×600-px-medium.jpg" alt="" />
+</div>
