+++ src/app/views/layout/sidebar-menu/sidebar-menu.component.html
@@ -1,1 +1,15 @@
-<p>sidebar-menu works!</p>
+<aside class="sideBar">
+  <ul class="nav nav-tabs" id="myTab" role="tablist">
+    <li class="nav-item" role="presentation">
+      <a class="nav-link active" id="tabPart_1" data-bs-toggle="tab" data-bs-target="#tab_1" href="#" role="tab"
+        aria-controls="tab_1" aria-selected="true">
+        <h4>JavaScript Quiz</h4>
+        <hr class="d-none d-md-block my-0 ms-0 me-0" />
+      </a>
+    </li>
+    <li class="nav-item" role="presentation" *ngFor="let quizlist2 of menu_sidebar">
+      <a class="nav-link" id="tabPart_2" data-bs-toggle="tab" data-bs-target="#tab_2" href="#" role="tab"
+        aria-controls="tab_2" aria-selected="false">{{ quizlist2?.name }}</a>
+    </li>
+  </ul>
+</aside>
