+++ src/app/views/layout/sidebar-menu/sidebar-menu.component.ts
@@ -1,4 +1,6 @@
 import { Component } from '@angular/core';
+import { CommonService } from '../../../common/common.service';
+import { ActivatedRoute, Router } from '@angular/router';
 
 @Component({
   selector: 'app-sidebar-menu',
@@ -6,5 +8,26 @@ import { Component } from '@angular/core';
   styleUrls: ['./sidebar-menu.component.css']
 })
 export class SidebarMenuComponent {
+  constructor(
+    public commonservice: CommonService,
+    private router: ActivatedRoute
+  ) { }
+
+  menu_sidebar: any;
+
+  ngOnInit() {
+    this.getMenuSidebar();
+  }
+  async getMenuSidebar() {
+    await this.commonservice
+      .get(`page/get-quiz-list/${this.router.snapshot.params['topic']}`)
+      .subscribe((result) => {
+        const apiResult2 = JSON.parse(JSON.stringify(result));
+
+        if (apiResult2 && apiResult2.status == 'SUCCESS') {
+          this.menu_sidebar = apiResult2 && apiResult2.payload;
+        }
+      });
+  }
 
 }
